+++ src/main/java/com/payflow/exception/InvalidUpiException.java
@@ -1,0 +1,8 @@
+package com.payflow.exception;
+
+public class InvalidUpiException extends PayflowException {
+
+	public InvalidUpiException(String message) {
+		super(message);
+	}
+}